+++ AlexanderMandrov/Lesson15/index.html
@@ -37,6 +37,7 @@
 
                 <div class="modal__content">
                     <div class="modal__name">name</div>
+                    <div class="modal__order"></div>
                     <div class="modal__email"></div>
                     <div class="modal__phone"></div>
                     <div class="modal__price">500 RUB</div>
@@ -50,7 +51,7 @@
             <ul class="list"></ul>
             <div class="total__price">Basket is clear</div>
         </div>
-        <canvas class="canvas" width="800" height="300"></canvas>
+        <canvas id="canvas" width="800" height="300"></canvas>
         <form action="#" class="form">
             <input type="text" class="name" placeholder="name"><br>
             <input type="text" class="email" placeholder="email"><br>
@@ -60,7 +61,10 @@
             <button class="order" type="button">Order</button>
         </form>
     </div>
-    <script src="utils.js" type="module"></script>
-    <script src="shop.js" type="module"></script>
+    <script src="./js/utils.js" type="module"></script>
+    <script src="./js/canvas.js" type="module"></script>
+    <script src="./js/services.js" type="module"></script>
+    <script src="./js/Basket.js" type="module"></script>
+    <script src="./js/shop.js" type="module"></script>
 </body>
 </html>
