+++ AlexanderMandrov/Lesson15/js/shop.js
@@ -1,258 +1,16 @@
 import isCorrectInput from './utils.js';
+import Basket from './Basket.js';
+import ItemsList from './ItemsList.js'
+import initCanvasDrawing from  './canvas.js';
 
-class Basket {
-    constructor(items, prices, currency, quantities) {
-        this.items = [...items];
-        this.prices = [...prices];
-        this.currency = currency;
-        this.quantities = [...quantities];
-    }
-
-    countBasketPrice() {
-        let sum = 0;
-        this.prices.forEach((price, i) => {
-            sum += price * this.quantities[i];
-        });
-        return sum;
-    }
-
-    getItem(i) {
-        return this.items[i];
-    }
-
-    getItems() {
-        return this.items;
-    }
-
-    getItemQuantity(i){
-        return +this.quantities[i];
-    }
-
-    getItemPrice(i) {
-        return this.prices[i];
-    }
-
-    getItemIndex(name) {
-        let j = 0;
-        [...this.getItems()].forEach((item, i) => {
-            if (item === name) j = i;
-        });
-        return j;
-    }
-
-    editItemQuantity(name, quantity) {
-        this.quantities[this.getItemIndex(name)] = quantity;
-    }
-
-    pushItem(item) {
-        this.items.push(item.name);
-        this.prices.push(item.price);
-        this.quantities.push(item.quantity);
-    }
-
-    deleteItem(i) {
-        this.items.splice(i, 1);
-        this.prices.splice(i, 1);
-        this.quantities.splice(i, 1);
-    }
-    
-    getOrder() {
-        let orderMessage = '';
-        [...this.getItems()].forEach((item, i) => {
-            orderMessage += `${item} ${this.getItemQuantity(i)}x, `
-        });
-
-        return orderMessage.split(0, orderMessage.length - 1);
-    }
-}
-
-let shopList = document.querySelector('.shop__list');
 let products = new Basket([], [], 'RUB', []);
-var counter = 0;
+const shopList = document.querySelector('.shop__list');
 
-const makeGETRequest = (url) => {
-    return new Promise((resolve, reject) => {
-        const xhr = new XMLHttpRequest();
+//draw charts
+initCanvasDrawing();
 
-        xhr.onreadystatechange = function () {
-            if (xhr.readyState === XMLHttpRequest.DONE) {
-                if (xhr.status != 200) {
-                    reject(`${xhr.status}: ${xhr.statusText}`);
-                }
-                resolve(JSON.parse(xhr.responseText));
-            }
-        }
 
-        xhr.open('GET', url, true);
-        xhr.send();
-    });
-}
 
-const eventHandler = () => {
-    const shopImages = document.querySelectorAll('.shop__img');
-    shopImages.forEach(img => {
-        img.addEventListener('click', (event) => {
-            const target = event.target;
-            const item = target.closest('.shop__item'),
-                counter = item.dataset.counter,
-                itemName = document.querySelectorAll('.shop__description')[counter].textContent,
-                itemPrice = document.querySelectorAll('.shop__price')[counter].textContent;
-    
-            modalName.textContent = itemName;
-            modalPrice.textContent = itemPrice;
-            modal.style.display = 'block';
-        });
-    });
-};
-
-//GET items
-const API_URL = '';
-
-const c = document.querySelector('.canvas');
-
-const getSalesData = async () => {
-    try {
-        const rawResponse = await fetch(`${API_URL}/sales`);
-        const response = rawResponse.json();
-        return response;
-    } catch {
-        return false;
-    }
-};
-
-const catchData = async () => {
-    const data = await getSalesData();
-    if (data) {
-        const salesCount = [],
-            monthArr = [],
-            newArr = [];
-
-        data.map((sale) => {
-            salesCount.push(sale.sales);
-            monthArr.push(sale.month);
-        });
-        const totalSales = salesCount.reduce((a, b) => a + b, 0);
-
-        salesCount.forEach(item => {
-            newArr.push(item / totalSales);
-        });
-
-        const colors = ["red", "blue", "green", "orange", "teal", "purple"];
-        const circleOptions = {
-            data: newArr,
-            month: monthArr,
-            colors: colors
-        };
-        const barOptions = {
-            data: salesCount,
-            month: monthArr,
-            colors: colors
-        };
-        makeCircle(circleOptions);
-        makeBar(barOptions);
-    }
-}
-
-const makeCircle = ({ data, month, colors }) => {
-    const ctx = c.getContext("2d");
-    ctx.font = "18px Georgia";
-    ctx.beginPath();
-    let initialPoint = 0;
-    let endPoint = data[0];
-    const xCenter = 105,
-        margin = 30,
-        radius = 100,
-        yCenter = 105;
-    ctx.lineTo(xCenter, yCenter);
-    for (let i = 0; i < data.length; i++){
-        ctx.beginPath();
-        ctx.arc(xCenter, yCenter, radius, initialPoint * 2 * Math.PI, endPoint * 2 * Math.PI);
-        if (i === data.length - 1) {
-            ctx.arc(xCenter, yCenter, radius, initialPoint * 2 * Math.PI, 0);
-        }
-        ctx.lineTo(xCenter, yCenter);
-        ctx.closePath();
-        ctx.fillStyle = colors[i];
-        ctx.strokeStyle = colors[i];
-        ctx.fillText(month[i], 250, margin * (i + 1));
-        ctx.fill();
-        initialPoint = endPoint;
-        endPoint = initialPoint + data[i + 1];
-    }
-    ctx.stroke();
-};
-
-const makeBar = ({ data, month, colors }) => {
-    const ctx = c.getContext("2d");
-    ctx.font = "11px Georgia";
-    let initialX = 400;
-    let initialY = 0;
-    let bottomLine = initialY + data[0];
-    ctx.moveTo(initialX, initialY);
-    ctx.beginPath();
-    for (let i = 0; i < data.length; i++) {
-        ctx.fillStyle = colors[i];
-        ctx.fillRect(initialX, bottomLine - data[i], 50, data[i]);
-        ctx.fillText(month[i], initialX, bottomLine + 10);
-        initialY = data[i];
-        initialX += 50;
-    };
-    ctx.stroke();
-};
-
-catchData();
-
-class ItemsList {
-    fetchItems () {
-        makeGETRequest(`${API_URL}/items`)
-            .then(items => {
-                            items.forEach(item => this.renderShopItem(item.name, item.price, item.currency));
-                            eventHandler();
-                        })
-            .catch(errMessage => console.log(errMessage));
-    }
-
-    renderShopItem(name, price, currency) {
-        const shopItem = document.createElement('div');
-        shopItem.classList.add('shop__item');
-        shopItem.dataset['counter'] = counter;
-        
-        const dataShopItem = `
-        <img src="https://placehold.it/200" alt="photo" class="shop__img">
-        <hr>
-        `;
-    
-        shopItem.insertAdjacentHTML('afterbegin', dataShopItem);
-    
-        const shopContent = document.createElement('div');
-        shopContent.classList.add('shop__content');
-        shopContent.dataset['counter'] = counter;
-    
-        const dataShopContent = `
-        <div class="shop__description" data-counter="${counter}">${name}</div>
-        <div class="shop__price" data-counter="${counter}">${price} ${currency}</div>
-        <input type="text" class="shop__input" value="1" data-counter="${counter++}"><br>
-        `;
-    
-        shopContent.insertAdjacentHTML('afterbegin', dataShopContent);
-    
-        const btns = document.createElement('div');
-        btns.classList.add('btns');
-    
-        const dataBtns = `
-        <button type="button" class="shop__btn">Buy</button>
-        <button type="button" class="shop__btn btn-delete">Delete</button>
-        `;
-    
-        btns.insertAdjacentHTML('afterbegin', dataBtns);
-    
-        shopContent.insertAdjacentElement('beforeend', btns);
-    
-        shopItem.insertAdjacentElement('beforeend', shopContent);
-        
-        shopList.insertAdjacentElement('beforeend', shopItem);
-    }
-}
 
 const goods = new ItemsList();
 goods.fetchItems();
@@ -430,6 +188,7 @@ const modal = document.querySelector('.modal'),
     modalComment = document.querySelector('.modal__comment'),
     modalEmail = document.querySelector('.modal__email'),
     modalPhone = document.querySelector('.modal__phone'),
+    modalOrder = document.querySelector('.modal__order'),
     modalAddress = document.querySelector('.modal__address');
 
 const orderBtn = document.querySelector('.order');
@@ -453,7 +212,7 @@ orderBtn.addEventListener('click', (event) => {
         }
     });
     if (flg) {
-        modalName.textContent = `You ordered: ${products.getOrder()}`;
+        modalOrder.textContent = `You ordered: ${products.getOrder()}`;
         modalPrice.textContent = `Total amount is ${products.countBasketPrice()} ${products.currency},`;
         modalComment.textContent = `Your comment: ${comment.value}`;
         modalPhone.textContent = `Your phone: ${phone.value},`;
@@ -471,7 +230,7 @@ document.addEventListener('click', e => {
     if (target.classList.contains('modal')) modal.style.display = 'none';
 });
 
-var mySwiper = new Swiper('.swiper-container', {
+const mySwiper = new Swiper('.swiper-container', {
     // Optional parameters
     direction: 'horizontal',
     loop: true,
