+++ AlexanderMandrov/Lesson15/js/utils.js
